encoder: config, DPB and GOP corrections
The parts of the encode path that a host exercises and the demo does not,
separated from the session object itself because none of them touch the
interface the ext layer binds to -- they compile and link against the
existing one.

Config gains the per-codec derivations a caller can reach directly rather
than through the demo's single fixed path. The H.264 DPB and the GOP
structure each carried state that was only ever read back in the order the
demo produced it; both now survive a caller that reconfigures mid-stream,
submits frames out of order, or stops early.

PSNR belongs to this topic by subject but not by dependency: it reads
fields that the next commit adds to the session, so it cannot compile until
that lands and it travels with it instead. Splitting on what builds rather
than on what reads well is deliberate -- a commit that does not compile is
not a reviewable unit.

Kept ahead of the session change so that the reviewable, self-contained
half of the core work is not buried inside the atomic commit that follows.

FINALIZECONFIG() TAKES AN OPTIONAL DEVICE CAPABILITY ARGUMENT, because some of
what it decides is a property of the hardware and not of the command line.

The function runs on two paths: argv parsing, which happens before any device
exists, and an embedding host, which has already probed one. A pointer that may
be null carries that difference without splitting the tail in two -- null keeps
the command-line answer, which is what the demo gets, and a caller that has
probed reaches a device-correct answer through the same function. A second
FinalizeAgainstDevice() would be a call every caller has to remember.

One decision uses it. Intra refresh is a device feature
(VkPhysicalDeviceVideoEncodeIntraRefreshFeaturesKHR::videoEncodeIntraRefresh), so
a configuration asking for it on a device that lacks it is refused here, where the
request is still attributable, rather than inside session creation.

The minQp default is deliberately NOT a second one. Every consumer of that field
reads it only under minQpSet, which stays clear precisely when the default
applies, so an unset minQp reaches the driver as "no clamp" rather than as 20 --
checking the number against the device would decide nothing. A caller that wants
a QP clamp sets one, and the codec configs check THAT against the device window
and refuse it rather than narrowing a request silently. DeviceCapabilities
carries only what it is asked, and defaults its members, so a field the caller
does not assign reads as absent rather than as whatever the stack held.

--GOPFRAMECOUNT IS PARSED AT THE WIDTH IT IS STORED. VkVideoGopStructure keeps
the count in a uint32_t, and its constructor now agrees with the member and the
setter. Parsing into a narrower local truncated silently, because parseUint casts
without a range check: --gopFrameCount 300 reached the encoder as 44, and 256 as
0, which the codec configs read as ZERO_GOP_FRAME_COUNT and replaced with the
device's preferred count. Neither is refused and neither is reported, so the
stream is encoded to a GOP the caller never asked for.

Zero remains legal: it is that sentinel, and requesting the device's preference
is a request like any other.

codecBlockAlignment is deliberately NOT among them. It holds the H.264 macroblock
size for every codec, which is wrong for H.265 CTBs and AV1 superblocks -- but
nothing in the tree reads it: one write here, one declaration, one constructor
initialiser and no readers. Giving it a device-derived value would add correctness
to a field no consumer checks, so it keeps its value and gains a comment saying
that a reader must derive the real granularity from
VkVideoCapabilitiesKHR::pictureAccessGranularity instead.

+++ vk_video_encoder/libs/VkVideoEncoder/VkVideoGopStructure.cpp
@@ -17,7 +17,7 @@
 #include "VkVideoGopStructure.h"
 #include <limits>
 
-VkVideoGopStructure::VkVideoGopStructure(uint8_t gopFrameCount,
+VkVideoGopStructure::VkVideoGopStructure(uint32_t gopFrameCount,
                                          int32_t idrPeriod,
                                          uint8_t consecutiveBFrameCount,
                                          uint8_t temporalLayerCount,
